+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gopowerteam/form-render
 
+## 0.0.11
+
+### Patch Changes
+
+- add table form-options props
+
 ## 0.0.10
 
 ### Patch Changes